Only allow install on Ubuntu 18.04 / Assign random SSH port

+++ amcfwm.sh
@@ -10,7 +10,7 @@
 #                                                                                                            #
 #                           AsusWRT-Merlin CFW Manager For Ubuntu 18.04 LTS                                  #
 #                            By Adamm - https://github.com/Adamm00/am_cfwm                                   #
-#                                         30/03/2020 - v1.0.0                                                #
+#                                         31/03/2020 - v1.0.0                                                #
 ##############################################################################################################
 
 
@@ -1092,40 +1092,45 @@ case "$1" in
 	;;
 
 	install)
-		sudo apt-get update
-		sudo apt-get -y dist-upgrade
-		sudo dpkg --add-architecture i386
-		sudo apt-get update
-		sudo apt-get -y install lib32ncurses5-dev dos2unix libtool-bin cmake libproxy-dev uuid-dev liblzo2-dev autoconf automake bash bison bzip2 diffutils file flex m4 g++ gawk groff-base libncurses5-dev libtool libslang2 make patch perl pkg-config shtool subversion tar texinfo zlib1g zlib1g-dev git gettext libexpat1-dev libssl-dev cvs gperf unzip python libxml-parser-perl gcc-multilib gconf-editor libxml2-dev g++-multilib gitk libncurses5 mtd-utils libncurses5-dev libvorbis-dev git autopoint autogen sed build-essential intltool libelf1 libglib2.0-dev xutils-dev lib32z1-dev lib32stdc++6 xsltproc gtk-doc-tools libelf-dev:i386 libelf1:i386 libltdl-dev openssh-server curl git build-essential nano
-		if [ ! -f "$HOME/Desktop/amcfwm.sh" ]; then curl -fsL --retry 3 "https://raw.githubusercontent.com/Adamm00/am_cfwm/master/amcfwm.sh" -o "$HOME/amcfwm/amcfwm.sh"; fi
-		sudo ln -sf "$HOME/amcfwm/amcfwm.sh" /bin/amcfwm
-		chmod 755 /bin/amcfwm
-		mkdir -p ~/images
-		echo
-		if [ ! -f ~/.ssh/id_rsa ]; then
-			ssh-keygen -b 4096
+		if grep -qF "Ubuntu 18.04" "/etc/os-release"; then
+			sudo apt-get update
+			sudo apt-get -y dist-upgrade
+			sudo dpkg --add-architecture i386
+			sudo apt-get update
+			sudo apt-get -y install lib32ncurses5-dev dos2unix libtool-bin cmake libproxy-dev uuid-dev liblzo2-dev autoconf automake bash bison bzip2 diffutils file flex m4 g++ gawk groff-base libncurses5-dev libtool libslang2 make patch perl pkg-config shtool subversion tar texinfo zlib1g zlib1g-dev git gettext libexpat1-dev libssl-dev cvs gperf unzip python libxml-parser-perl gcc-multilib gconf-editor libxml2-dev g++-multilib gitk libncurses5 mtd-utils libncurses5-dev libvorbis-dev git autopoint autogen sed build-essential intltool libelf1 libglib2.0-dev xutils-dev lib32z1-dev lib32stdc++6 xsltproc gtk-doc-tools libelf-dev:i386 libelf1:i386 libltdl-dev openssh-server curl git build-essential nano
+			if [ ! -f "$HOME/Desktop/amcfwm.sh" ]; then curl -fsL --retry 3 "https://raw.githubusercontent.com/Adamm00/am_cfwm/master/amcfwm.sh" -o "$HOME/amcfwm/amcfwm.sh"; fi
+			sudo ln -sf "$HOME/amcfwm/amcfwm.sh" /bin/amcfwm
+			chmod 755 /bin/amcfwm
+			mkdir -p "$HOME/images"
+			echo
+			if [ ! -f "$HOME/.ssh/id_rsa" ]; then
+				ssh-keygen -b 4096
+			fi
+			echo "Your SSH Pubkey For Remote SSH Access - Save This File - [Press Enter To Continue]"
+			cat "$HOME/.ssh/id_rsa.pub"
+			read -r
+			echo "Setting Up OpenSSH-Server - Input authorized_keys Of Storage Server - [Press Enter To Continue]"
+			read -r
+			sudo nano -w "$HOME/.ssh/authorized_keys"
+			echo "Hardening OpenSSH Config"
+			SSHPORT="$(awk -v min=49152 -v max=65535 -v freq=1 'BEGIN{"tr -cd 0-9 </dev/urandom | head -c 6" | getline seed; srand(seed); for(i=0;i<freq;i++)print int(min+rand()*(max-min+1))}')"
+			sed -i "s~.*Port .*~Port $SSHPORT~g" /etc/ssh/sshd_config
+			echo "VM SSH Port Changed To $SSHPORT"
+			sed -i 's~#ChallengeResponseAuthentication yes~ChallengeResponseAuthentication no~g' /etc/ssh/sshd_config
+			sed -i 's~#PasswordAuthentication yes~PasswordAuthentication no~g' /etc/ssh/sshd_config
+			echo "SSH Password Authentication Disabled"
+			echo "Input MOTD - [Press Enter To Continue]"
+			read -r
+			rm -rf /etc/update-motd.d/80-livepatch /etc/update-motd.d/50-motd-news /etc/update-motd.d/80-esm /etc/update-motd.d/91-release-upgrade /etc/update-motd.d/95-hwe-eol
+			true > /etc/update-motd.d/10-help-text
+			sudo nano -w /etc/update-motd.d/10-help-text
+			Set_Default
+			echo "Rebooting To Apply Updates - [Press Enter To Continue]"
+			read -r
+			sudo rm -f /bin/sh && sudo ln -sf bash /bin/sh && sudo reboot
+		else
+			echo "OS Not Supported - Please Install Ubuntu 18.04 LTS To Use This Script"
 		fi
-		echo "Your SSH Pubkey For Remote Access - [Press Enter To Continue]"
-		cat ~/.ssh/id_rsa.pub
-		read -r
-		echo "Setting Up OpenSSH-Server - Input authorized_keys - [Press Enter To Continue]"
-		read -r
-		sudo nano -w ~/.ssh/authorized_keys
-		echo "Hardening OpenSSH Config"
-		echo "Password Authentication Disabled"
-		echo "SSH Port Changed To 4216"
-		sed -i 's~#Port 22~Port 4216~g' /etc/ssh/sshd_config
-		sed -i 's~#ChallengeResponseAuthentication yes~ChallengeResponseAuthentication no~g' /etc/ssh/sshd_config
-		sed -i 's~#PasswordAuthentication yes~PasswordAuthentication no~g' /etc/ssh/sshd_config
-		echo "Input MOTD - [Press Enter To Continue]"
-		read -r
-		rm -rf /etc/update-motd.d/80-livepatch /etc/update-motd.d/50-motd-news /etc/update-motd.d/80-esm /etc/update-motd.d/91-release-upgrade /etc/update-motd.d/95-hwe-eol
-		true > /etc/update-motd.d/10-help-text
-		sudo nano -w /etc/update-motd.d/10-help-text
-		Set_Default
-		echo "Rebooting To Apply Updates - [Press Enter To Continue]"
-		read -r
-		sudo rm -f /bin/sh && sudo ln -sf bash /bin/sh && sudo reboot
 	;;
 
 	repo)
@@ -1133,34 +1138,34 @@ case "$1" in
 		sudo mkdir -p /opt
 
 		cd ~ || exit 1
-		if [ ! -d ~/am-toolchains ]; then
+		if [ ! -d "$HOME/am-toolchains" ]; then
 			echo "Preparing Toolchain Repo"
 			git clone https://github.com/RMerl/am-toolchains.git
-			sudo ln -s ~/am-toolchains/brcm-arm-hnd /opt/toolchains
+			sudo ln -s "$HOME/am-toolchains/brcm-arm-hnd" "/opt/toolchains"
 		fi
-		if [ ! -d ~/amng ]; then
+		if [ ! -d "$HOME/amng" ]; then
 			echo "Preparing Firmware Repo"
 			git clone https://github.com/RMerl/asuswrt-merlin.ng amng
 		fi
 
-		sed -i '\~AsusWRT-Merlin CFW Manager~d' ~/.profile
+		sed -i '\~AsusWRT-Merlin CFW Manager~d' "$HOME/.profile"
 
 		# BCM-SDK
-		sudo ln -fs ~/am-toolchains/brcm-arm-sdk/hndtools-arm-linux-2.6.36-uclibc-4.5.3 /opt/brcm-arm
-		sudo ln -fs ~/am-toolchains/brcm-arm-sdk  ~/amng/release/src-rt-6.x.4708/toolchains
-		echo "PATH=$PATH:/opt/brcm-arm/bin # AsusWRT-Merlin CFW Manager" >> ~/.profile
+		sudo ln -fs "$HOME/am-toolchains/brcm-arm-sdk/hndtools-arm-linux-2.6.36-uclibc-4.5.3" "/opt/brcm-arm"
+		sudo ln -fs "$HOME/am-toolchains/brcm-arm-sdk"  "$HOME/amng/release/src-rt-6.x.4708/toolchains"
+		echo "PATH=$PATH:/opt/brcm-arm/bin # AsusWRT-Merlin CFW Manager" >> "$HOME/.profile"
 
 		# BCM-HND
-		sudo ln -fs ~/am-toolchains/brcm-arm-hnd /opt/toolchains
+		sudo ln -fs "$HOME/am-toolchains/brcm-arm-hnd" "/opt/toolchains"
 		{ echo "export LD_LIBRARY_PATH=$LD_LIBRARY:/opt/toolchains/crosstools-arm-gcc-5.3-linux-4.1-glibc-2.22-binutils-2.25/usr/lib # AsusWRT-Merlin CFW Manager"
 		echo "export TOOLCHAIN_BASE=/opt/toolchains # AsusWRT-Merlin CFW Manager"
 		echo "PATH=\$PATH:/opt/toolchains/crosstools-arm-gcc-5.3-linux-4.1-glibc-2.22-binutils-2.25/usr/bin # AsusWRT-Merlin CFW Manager"
-		echo "PATH=\$PATH:/opt/toolchains/crosstools-aarch64-gcc-5.3-linux-4.1-glibc-2.22-binutils-2.25/usr/bin # AsusWRT-Merlin CFW Manager"; } >> ~/.profile
+		echo "PATH=\$PATH:/opt/toolchains/crosstools-aarch64-gcc-5.3-linux-4.1-glibc-2.22-binutils-2.25/usr/bin # AsusWRT-Merlin CFW Manager"; } >> "$HOME/.profile"
 
 		# BCM-HND AX
 		{ echo "export LD_LIBRARY_PATH=$LD_LIBRARY:/opt/toolchains/crosstools-arm-gcc-5.5-linux-4.1-glibc-2.26-binutils-2.28.1/usr/lib # AsusWRT-Merlin CFW Manager"
 		echo "PATH=\$PATH:/opt/toolchains/crosstools-arm-gcc-5.5-linux-4.1-glibc-2.26-binutils-2.28.1/usr/bin # AsusWRT-Merlin CFW Manager"
-		echo "PATH=\$PATH:/opt/toolchains/crosstools-aarch64-gcc-5.5-linux-4.1-glibc-2.26-binutils-2.28.1/usr/bin # AsusWRT-Merlin CFW Manager"; } >> ~/.profile
+		echo "PATH=\$PATH:/opt/toolchains/crosstools-aarch64-gcc-5.5-linux-4.1-glibc-2.26-binutils-2.28.1/usr/bin # AsusWRT-Merlin CFW Manager"; } >> "$HOME/.profile"
 
 		echo
 		echo "Repo Setup Complete!"
@@ -1596,9 +1601,9 @@ case "$1" in
 			case "$continue" in
 				1)
 					sudo rm -rf "$HOME/am-toolchains" "$HOME/amng"  "$HOME"/amng.* "/opt/toolchains" "/opt/brcm-arm"
-					sed -i '\~AsusWRT-Merlin CFW Manager~d' ~/.profile
+					sed -i '\~AsusWRT-Merlin CFW Manager~d' "$HOME/.profile"
 					sudo rm -rf "$HOME/amcfwm" "/bin/amcfwm"
-					sudo rm -rf /etc/update-motd.d/10-help-text
+					sudo rm -rf "/etc/update-motd.d/10-help-text"
 				;;
 				2|e|exit)
 					echo "[*] Exiting!"
